requirements.txt
Flask==3.0.0
gunicorn==21.2.0
tensorflow-cpu==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
numpy==1.23.5
Pillow==9.5.0
h5py==3.8.0
werkzeug==3.0.1
